more pages less drawers --> more space
